# docs/swagger.yaml
basePath: /
definitions:
  model.CheckRequest:
    properties:
      key:
        type: string
      rule_id:
        type: integer
    required:
    - key
    - rule_id
    type: object
  model.CheckResponse:
    properties:
      allowed:
        type: boolean
    type: object
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  model.Rule:
    properties:
      created_at:
        type: string
      id:
        type: integer
      limit_count:
        type: integer
      name:
        type: string
      window_seconds:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Distributed Rate Limiter using Go, Redis and PostgreSQL
  title: Distributed Rate Limiter API
  version: "1.0"
paths:
  /check:
    post:
      consumes:
      - application/json
      description: Evaluates whether a request is allowed
      parameters:
      - description: Rate limit check request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CheckResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Check rate limit
      tags:
      - Rate Limiter
  /rules:
    get:
      description: Returns all configured rate limit rules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Rule'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: List all rules
      tags:
      - Rules
    post:
      consumes:
      - application/json
      description: Create a new rate limiting rule
      parameters:
      - description: Rule payload
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/model.Rule'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Rule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create rate limit rule
      tags:
      - Rules
  /rules/{id}:
    get:
      description: Fetch a rule by ID
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get rule by ID
      tags:
      - Rules
swagger: "2.0"